% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machuruku_v0.1.8.3.R
\name{machu.1.tip.resp}
\alias{machu.1.tip.resp}
\title{Calculate tip climate response curves}
\usage{
machu.1.tip.resp(
  occ,
  ClimCur,
  sp.col = 1,
  long.col = 2,
  lat.col = 3,
  verbose = FALSE
)
}
\arguments{
\item{occ}{occurrence data for each species, formatted as a dataframe}

\item{ClimCur}{present-day climate data, formatted as a RasterStack or RasterBrick}

\item{sp.col}{specify which column of the input occurrence data corresponds to species ID (1 by default)}

\item{long.col}{specify longitude column (2 by default)}

\item{lat.col}{specify latitude column (3 by default)}

\item{verbose}{if TRUE, print progress (by-taxon) to the screen}
}
\value{
table consisting of the response of each species to the climate
data. Each response is represented by a skew-normal distribution.
}
\description{
For each taxon, construct a BIOCLIM species distribution model
and estimate a response curve for each climate variable.
}
\examples{
response.table <- machu.1.tip.resp(occ, ClimCur)
}
