% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machuruku_v0.1.8.3.R
\name{machu.2.ace}
\alias{machu.2.ace}
\title{Perform ancestral character estimation of response curves}
\usage{
machu.2.ace(tip.resp, tree, T = NULL, n.unc = NULL, savename = NULL)
}
\arguments{
\item{tip.resp}{a table containing the character data for each tip.
Corresponds to the output of machu.1.tip.resp().}

\item{tree}{an ultrametric phylogeny, formatted as a tree object. Tip
labels must be identical to those in the tip.resp table. Note that
treedata objects do not work (use as.phylo() to convert).}

\item{T}{time from which to extract reconstructed values along any
branches present at T. Must be a value between 0 and the root age of
the tree. Should correspond to the estimated date of paleoclimatic
layers for best results. If T is not provided, the function will output
climate response curves for each node (as well as each tip).}

\item{n.unc}{a whole, positive number that specifies the number of evenly-spaced
samples of the 95\% confidence interval to take for characterizing uncertainty in
ancestral character estimations. n.unc sampled values will be tacked on to the end
of the output, and in machu.3.ace(), specifying calc.unc = T will construct separate
models for each sample and average them to get the final model.}

\item{savename}{a string that, when provided, saves the output as a .csv file. It is not
necessary to add the file extension to the string. Use machu.ace.load() to load one of these
saved files.}
}
\value{
a list where each element corresponds to a response curve parameter for a
certain climate variable, and contains the values of that parameter for each taxon
in question. If T is not provided (default), the taxa returned are the tips and nodes
of the provided tree. If T is provided, the taxa returned are the branches present
at time T. If n.unc is provided, n.unc additional models will also be generated
based on the 95\% confidence intervals of each value. These additional models will
be used in step 3 to incorporate uncertainty in ancestral character estimation
into the process by generating an ensemble model.
}
\description{
Use ace() to generate climate response curves at each node of a
time-calibrated phylogeny, and extract the values along the branches
at a particular time if so desired. By default, the output is a set
of climate response curves for each node (and each tip). As each node
occurs at different times (making paleoclimatic projections tricky),
the user can also provide a time-slice value T. The function will
find each branch present at time T and interpolate response curves
along them to that point, then return those values. The function can
also estimate uncertainty with the n.unc parameter.
}
\examples{
# get model for each node, without uncertainty
ace.out <- machu.2.ace(response.table, bassleritree)
# get model for each lineage at T=2.5, without uncertainty
ace.out <- machu.2.ace(response.table, bassleritree, T=2.5)
# get model for each node, with 4 uncertainty samples
ace.out <- machu.2.ace(response.table, bassleritree, n.unc=4)
# get model for each lineage at T=2.5, with 4 uncertainty samples
ace.out <- machu.2.ace(response.table, bassleritree, T=2.5, n.unc=4)
}
