% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machuruku_v0.1.8.2.R
\name{machu.3.anc.niche}
\alias{machu.3.anc.niche}
\title{Convert ancestral response values to niche models}
\usage{
machu.3.anc.niche(
  ace,
  clim,
  taxa = NULL,
  resp.curv = TRUE,
  clip.Q = TRUE,
  calc.unc = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{ace}{a list of tables with the reconstructed response values
for each ancestor at T. Corresponds to the output of machu.2.ace().}

\item{clim}{paleoclimatic data, formatted as a RasterStack. Should
be as close to time T as possible for best results.}

\item{taxa}{optionally select specific taxa from ace to create models
for. This is useful if you wish to visualize only one specific taxon or
node at a certain time. The form is a row index or vector of indices. For
example, taxa = 1 will only create a model for the first taxon, and
taxa = 1:3 will create models for the first three. taxa = c(1,3:5) will
create models for the first taxon and for taxa 3 to 5.}

\item{resp.curv}{if TRUE, create ancestral models using response curves.
If false, assume response curve is a uniform distribution between the LCI
and UCI response values. Default = TRUE. Setting resp.curv = FALSE will
also automatically set clip.Q = FALSE.}

\item{clip.Q}{if TRUE, clip the tails of each response curve at .025 and
0.975. May produce cleaner models. Default = TRUE.}

\item{calc.unc}{if TRUE, incorporate uncertainty bins from using n.unc
in machu.2.ace(). Visualizes uncertainty in the model. Default = FALSE.}

\item{verbose}{if TRUE, print to screen the current taxon and climate
layer being processed. Default = FALSE.}
}
\value{
a list consisting of one raster per ancestor at T, showing the
habitat suitability for each pixel on the landscape.
}
\description{
Create a suitability map for each ancestor at time T, then convert
into a Bioclim model for each ancestor.
}
\examples{
#run w/ default settings (resp.curv & clip.Q = TRUE, do not account for uncertainty)
output.models <- machu.3.anc.niche(ace.output, clim)
#assume response curves are uniform distributions between min and max
output.models <- machu.3.anc.niche(ace.output, clim, resp.curv=F)
#Do not clip response curve tails
#Note that clip.Q cannot be true if resp.curv is false.
output.models <- machu.3.anc.niche(ace.output, clim, clip.Q=F)
#incorporate uncertainty (make sure to specify a value for n.unc in machu.2.ace() beforehand)
output.models <- machu.3.anc.niche(ace.out.N, ClimMis19, calc.unc = T)
}
