% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machuruku_v0.1.8.R
\name{machu.ace.load}
\alias{machu.ace.load}
\title{Load saved ancestral character estimations from .csv file}
\usage{
machu.ace.load(file)
}
\arguments{
\item{file}{filename (or filepath) of the .csv saved by machu.2.ace().}
}
\value{
correctly-formatted output of machu.2.ace().
}
\description{
Using the savename parameter in machu.2.ace(), you can save output
as a .csv. This function allows you to load this back into the special
list format which is necessary for machu.3.anc.niche().
}
\examples{
load.ace <- machu.ace.load("saved_ace.csv")

}
