% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machuruku_v0.1.8.R
\name{machu.geo.idw}
\alias{machu.geo.idw}
\title{Clip models via inverse-distance weighting}
\usage{
machu.geo.idw(
  in.sdm,
  in.pts,
  buffer.dist = 300,
  kernel.size = 2,
  MCP.percent = 100,
  taxa.col = 1,
  long.col = 2,
  lat.col = 3,
  taxa = NULL
)
}
\arguments{
\item{in.sdm}{niche model, a single raster object. Corresponds to one element
of the list output of machu.3.anc.niche().}

\item{in.pts}{occurrence points, a dataframe. Default input format is taxon, long, lat,
though this can be finagled with other arguments.}

\item{buffer.dist}{distance (in km) by which to buffer a minimum convex polygon (MCP)
around the occurrence points. Any model values inside this MCP will remain unchanged.}

\item{kernel.size}{a multiplier that defines the broadness of the smoothing edge
surrounding the MCP created by buffer.dist. A kernel.size of 2 creates a smoothing
edge twice the width of the MCP buffer.dist. Thus, higher values create broader models.
Minimum value of 1.}

\item{MCP.percent}{defines proportion of outlier occurrence points to be excluded from
the MCP creation.}

\item{taxa.col}{the column in the occurrence data with taxon names. Set to 1 by default.}

\item{long.col}{the column in the occurrence data with longitudes (x). Set to 2 by default.}

\item{lat.col}{the column in the occurrence data with latitudes (y). Set to 3 by default.}

\item{taxa}{a string or string vector declaring which taxa are to be used in the creation of the
MCP. Must match exactly. If the argument is not called, the function uses all occurrence
points by default.}
}
\value{
a clipped niche model, formatted as a raster layer.
}
\description{
Niche modeling may recover areas that are distant and non-contiguous to a species' core
range as suitable habitat. This can be either a feature or a bug. In certain cases one
may wish to exclude these "extraneous" suitable areas from an analysis because they could
not realistically be migrated to and are thus superfluous. This function uses inverse-distance
weighting to clip extraneous suitable areas a certain distance from a core range, as
delineated by occurrence data. Because occurrence data for extinct lineages does not exist
(barring fossils, which do not exist for many taxa), extant taxon occurrence data is used by
proxy. We suggest using machu.treeplot() to discern the extant descendants of the lineage
in question, and then restricting the occurrence data to those taxa using the taxa argument
of this function. Currently, the function only takes one niche model as input at a time.
}
\examples{
# after running through the "quick-start" Machuruku tutorial:
clip1<-machu.geo.idw(mod[[1]], occ, taxa=c("pepperi","bassleri","yoshina"), buffer.dist = 100, kernel.size = 2, MCP.percent = 50)
plot(mod[[1]]);plot(clip1)
# repeat for the silverstonei ancestor:
clip2<-machu.geo.idw(mod[[2]], occ, taxa="silverstonei", buffer.dist = 100, kernel.size = 2, MCP.percent = 50)
plot(mod[[2]]);plot(clip2)
}
