% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machuruku_code.R
\name{machu.occ.rarefy}
\alias{machu.occ.rarefy}
\title{Rarefy occurrence points}
\usage{
machu.occ.rarefy(
  in.pts,
  colxy = 2:3,
  rarefy.dist = 0,
  rarefy.units = "km",
  plot = F,
  verbose = T
)
}
\arguments{
\item{in.pts}{Input dataframe.}

\item{colxy}{Columns corresponding to longitude, then latitude. Default = 2:3.}

\item{rarefy.dist}{Distance to rarefy points (values need to be in km (recommended) or decimal degrees).  See associated parameter rarefy.units. Default = 0.}

\item{rarefy.units}{The units of the rarefy.dist parameter, either "km" for kilometers or "dd" for decimal degrees. Default = "km".}

\item{plot}{Creates an optional plot visualizing the points removed and kept. Default = F.}

\item{verbose}{If verbose=T, text boxes displaying progress will be displayed. Default = T.}
}
\value{
A dataframe with rarefied occurrence data.
}
\description{
Rarefy spatial occurrence points to reduce spatial autocorrelation and model bias. This is a modified version of humboldt.occ.rarefy() from humboldt.
}
\details{
A script to systematically select localities within a specified area at specified spatial resolution.  The outcome is always the same and is not random.  This reduces sampling biases in downstream analyses- you should do it! Output is a reduced dataset with less spatial autocorrelation.
}
\examples{
##remove occurrences closer than a minimum distance to each other (remove aggregation). Setting min.dist=0 will remove no occurrence.
occ <- machu.occ.rarefy(in.pts = occ, colxy = 2:3, rarefy.dist = 50, rarefy.units = "km")
}
