% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machuruku_v0.1.8.3.R
\name{machu.plotmap}
\alias{machu.plotmap}
\title{Map model output from machu.3.anc.niche()}
\usage{
machu.plotmap(model, col = 1, title = TRUE)
}
\arguments{
\item{model}{output from machu.3.anc.niche(). A list of rasters representing
the models to display. Using the full output displays each model successively.
User can also specify a subset of models using single-bracket notation to
only plot certain models or plot one model at a time.}

\item{col}{color ramp to use, a numeric value from 1 to 6. 1 = base, 2 =
plasma, 3 = viridis, 4 = l17, 5 = r3, 6 = white-to-black. Default = 1.}

\item{title}{if TRUE, print the taxon name above each corresponding model.
Default = TRUE.}
}
\value{
a series of maps corresponding to each output model specified.
}
\description{
Plot one or more maps using model output from machu.3.anc.niche().
User can specify color ramps from the viridis package.
}
\examples{
#plot all models with default color ramp and titles:
machu.plotmap(output.models)
#plot only the first model and use the "plasma" color ramp:
machu.plotmap(output.models[1], color=2)
#plot only the second and third models and do not display titles:
machu.plotmap(output.models[2:3], title=F)
#display three models at the same time
par(mfrow=c(1,3))
machu.plotmap(output.models)
#this function is designed for ease of use. For more control over plotting,
#use the image() function. Example below:
image(output.models[[1]], col=viridis(256), asp=1, axes=FALSE, xaxs="i", xaxt='n', yaxt='n', ann=F)
}
