% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machuruku_v0.1.8.3.R
\name{machu.top.env}
\alias{machu.top.env}
\title{Select top environmental variables}
\usage{
machu.top.env(
  occ,
  clim,
  sp.col = 1,
  long.col = 2,
  lat.col = 3,
  learning.rt = 0.01,
  pa.ratio = 4,
  steps = 50,
  method = "contrib",
  nvars.save = 5,
  contrib.greater = 5,
  verbose = T
)
}
\arguments{
\item{occ}{occurrence data for all taxa. Identical to input for machu.1.tip.resp(). Dataframe, with columns in the order of species, x/long, y/lat.}

\item{clim}{climate data for all taxa. Identical to input for machu.1.tip.resp(). A rasterstack of corresponding climate variables.}

\item{sp.col}{specify which column of the input occurrence data corresponds to species ID (1 by default)}

\item{long.col}{specify longitude column (2 by default)}

\item{lat.col}{specify latitude column (3 by default)}

\item{learning.rt}{value from 0.01 to 0.001 for building the ENMs, start with 0.01 and if prompted, change to 0.001. the default value is 0.01}

\item{pa.ratio}{ratio of pseudoabsences to occurrence points, typically this is 4. The default value is 4. There have to be at least 50 total points (occ+pseudoabsences) for the model to work; if the sum does not total to 50, the difference is taken as the number of pseudoabsences, rather than the value of occ*pa.ratio.}

\item{steps}{numbers of trees to add at each cycle for modelling each taxon. Start with 50 and if you run into problems gradually decrease, stopping at 1. The default value is 50}

\item{method}{this determines how important environmental variables are selected.There are three options: "estimate", "contrib", "nvars". If method="estimate", the boosted regression tree algorithm will choose the number of variables to include by systematically removing variables until average change in the model exceeds the original standard error of deviance explained. This is the most computationally intensive method. If method="contrib", variables above a relative influence value will be kept. See associated parameter 'contrib.greater'. If method="nvars", a fixed number of user specified variables will be kept. See associated parameter 'nvars.save'. The kept variables are selected by their relative influence. The 'nvars.save'-highest contributing variables for each taxon are retained and pooled, then ranked, and the 'nvars.save'-highest contributing variables for the whole pool are finally retained.}

\item{nvars.save}{if method="nvars",this variable is required. It is the number of the top variables to save. The kept variables are selected by their relative influence in predicting the species distribution, selecting for the highest contributing variables. Often the total variables retained is lower due to identical variables select among both species. The default value is 5.  This value will be ignored if method="estimate" or "contrib"}

\item{contrib.greater}{if method="contrib", this variable is required. The kept variables are selected for their relative influence in predicting the species' distribution. Here, users select variables equal to or above an input model contribution value. The default value for this method is 5 (= variables with 5 percent or higher contribution to model of either species are kept). This value will be ignored if method="estimate" or "nvars"}

\item{verbose}{print progress to the screen if TRUE, do not if FALSE}
}
\value{
prints the important variables to the screen. It is suggested to then combine them into a new rasterstack.
}
\description{
Perform a boosted regression tree analysis to identify the most important climate variables for your taxon set. This function is a modified version of humboldt.top.env() from humboldt. It runs generalized boosted regression models (a machine learning ENM algorithm) to select top parameters for inclusion your analyses. This is important because you want the models to reflect variables that are relevant to the species' distribution. Alternatively, you can run Maxent outside of R and manually curate the variables you include (also recommended).
}
\examples{
# identify the top 6 climate variables across all taxa
machu.top.env(occ, clim, method = "nvars", nvars.save = 6)
# identify all climate variables with a contribution greater than 10\%
machu.top.env(occ, clim, method = "contrib", contrib.greater = 10)
}
