% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machuruku_v0.1.8.2.R
\name{machu.trees.unc}
\alias{machu.trees.unc}
\title{Account for uncertainty in divergence time estimation from a posterior distribution of trees}
\usage{
machu.trees.unc(file, burnin = 0.1, conf = 0.95, inc = 10000, verbose = TRUE)
}
\arguments{
\item{file}{input file, provided as a filepath string. File must be NEXUS format with multiple
trees and a translation table of taxon names (e.g., 1 Homo sapiens). Only tested with BEAST 2 output.}

\item{burnin}{a numeric value between 0 and 1, denoting the percentage of trees to skip. MCMC chains do
not usually converge right away, so it is common practice to skip the first 10\% or so of samples as "burn-in".
The default value is 10\%.}

\item{conf}{a numeric value between 0 and 1, denoting the desired tree-height HPD interval to sample trees from. Values
closer to 1 will result in more extreme trees. The default value is 95\%.}

\item{inc}{the increment value at which the function will shout its progress at you, if verbose=TRUE. Default
is 10,000.}

\item{verbose}{shout progress, turned on by default}
}
\value{
returns a list of three phylo objects characterizing the uncertainty in divergence time of the posterior
distribution of trees and enabling the user to use the trees as separate inputs for Machuruku.
}
\description{
This function creates a list of three trees that characterize the uncertainty in divergence
times from a Bayesian posterior distribution of trees. It is written to work with a NEXUS file
containing multiple trees and a translation table of taxon names, specifically the output of
BEAST 2 but the format is generic enough to likely work with other program outputs as well.
The function is written to use a file without loading it into R and overloading memory, as
posterior treefiles can be very large in size.
The function conducts three passes through the posterior:
\enumerate{
\item Calculates the overall number of trees and the amount skipped as burn-in
\item Calculates the distribution of tree-heights and attendant 95\% HPD interval
\item Finds the trees with tree-heights closest to the upper 95\% HPD, lower 95\% HPD, and average
The output of the function is a list of three phylo objects, each of which can be separately
used as tree input for Machuruku.
}
}
\examples{
trees <- machu.trees.unc(file)
}
