% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machuruku_v0.1.8.2.R
\name{machu.1.tip.resp}
\alias{machu.1.tip.resp}
\title{Calculate tip climate response curves}
\usage{
machu.1.tip.resp(occ, ClimCur, verbose = FALSE)
}
\arguments{
\item{occ}{occurrence data for each species, formatted as a dataframe. Format should be species, x/long, and y/lat in that order.}

\item{ClimCur}{present-day climate data, formatted as a RasterStack}

\item{verbose}{if TRUE, print progress (by-taxon) to the screen}
}
\value{
table consisting of the response of each species to the climate
data. Each response is represented by a skew-normal distribution.
}
\description{
For each taxon, construct a BIOCLIM species distribution model
and estimate a response curve for each climate variable.
}
\examples{
response.table <- machu.1.tip.resp(occ, ClimCur)
}
