% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machuruku_v2.0.1.R
\name{machu.ace.load}
\alias{machu.ace.load}
\title{Load saved ace outputs from .csv file}
\usage{
machu.ace.load(file)
}
\arguments{
\item{file}{String specifying the file to load.}
}
\value{
A list of one element per scenario, either one or more timeslices or all tips and nodes, each a table of reconstructed climate response parameters.
}
\description{
Loads outputs saved with the 'csv.name' argument of machu.2.ace() into a format compatible with machu.3.anc.niche().
}
\details{
Because of the idiosyncrasies of the machu.2.ace() code, the result of machu.ace.load() is not technically identical to one obtained from the former function. Both should be compatible with machu.3.anc.niche(), but let me know if you encounter bugs.
}
\examples{
ace <- machu.ace.load("ace.csv")
}
