% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machuruku_v0.1.8.2.R
\name{machu.occ.rarefy}
\alias{machu.occ.rarefy}
\title{Rarefy occurrence points}
\usage{
machu.occ.rarefy(
  in.pts,
  colxy = 2:3,
  rarefy.dist = 0,
  rarefy.units = "km",
  verbose = T
)
}
\arguments{
\item{in.pts}{input data frame}

\item{colxy}{columns corresponding to longitude, then latitude}

\item{rarefy.dist}{distance to rarefy points (values need to be in km (recommended) or decimal degrees).  See associated parameter rarefy.units.}

\item{rarefy.units}{the units of rarefy.dist parameter, either "km" for kilometers or "dd" for decimal degrees}

\item{verbose}{if verbose=T, texts boxes displaying progress will be displayed}
}
\value{
A script to systematically select localities within a specified area at specified spatial resolution.  The outcome is always the same and is not random.  This reduces sampling biases in downstream analyses- you should do it! Output is a reduced dataset with less spatial autocorrelation.
}
\description{
Rarefy spatial occurrence points to reduce spatial autocorrelation and model bias. This is a modified version of humboldt.occ.rarefy() from humboldt.
}
\examples{
##remove occurrences closer than a minimum distance to each other (remove aggregation). Setting min.dist=0 will remove no occurrence.
occ <- machu.occ.rarefy(in.pts = occ, colxy = 2:3, rarefy.dist = 50, rarefy.units = "km")
}
