% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machuruku_v0.1.8.R
\name{machu.respplot}
\alias{machu.respplot}
\title{Visualize response curves of taxa to climate variables}
\usage{
machu.respplot(
  resp,
  taxa = NULL,
  clim = NULL,
  comb = FALSE,
  linetype = 1,
  linewidth = 1,
  linecolor = "black",
  legend.pos = "topleft",
  par.override = NULL
)
}
\arguments{
\item{resp}{input. This can either be output from machu.1.tip.resp() or from machu.2.ace().}

\item{taxa}{specify which taxa you wish to plot. This can either be done with text (e.g., "species1", or
c("species1,"species2")) or integers (e.g., 1 or 1:3, or c(1,4,6:10)). Note that with integers you must
refer to the "resp" input to know which integer corresponds to which taxon. When no taxa are specified,
all taxa are displayed by default.}

\item{clim}{specify which climate variables you wish to plot. This can only be done via text, and the
provided text must match the climate variable name exactly, e.g. "bio_1" or c("bio_1", "bio_10"). When
no climate variables are specified, all variables are displayed by default.}

\item{comb}{specify whether to combine taxon plots for the same climate variable. This option will
plot only as many charts as there are climate variables, and all taxa will be displayed in each chart,
allowing for more direct comparison. Different taxa are colored using the "Set2" color palette from
RColorBrewer. Color interpolation is used in cases of more than 8 taxa (the number of colors in the
palette). Default is set to FALSE.}

\item{linetype}{specify the desired line type for the plots. Default is set to 1, a solid line.}

\item{linewidth}{specify the desired line width for the plots. Default is set to 1.}

\item{linecolor}{specify the desired line color for the plots. Default is set to black. Note that this
only applies in the case where comb = FALSE. When comb = TRUE, the "Set2" color palette from RColorBrewer
is used.}

\item{legend.pos}{specify the desired legend position. You may wish to change this depending on the shape
of the curves, so as not to occlude them. The default setting is "topleft".}

\item{par.override}{override the default panel settings by providing a vector (e.g., c(2,3)) of the
desired configuration of rows and columns for the panels. By default, when comb = FALSE, each column
corresponds to a climate variable, and each row corresponds to a taxon; when comb = TRUE, all plots
are provided in a single row. For large numbers of climate variables, you may wish to specify multiple
rows of panels with this parameter.}
}
\value{
Returns a plot or plots showing the response of the specified taxa to the specified climate variable(s).
}
\description{
Visualize species response curves in a variety of different ways. A "response table" (the output
of machu.1.tip.resp()) or an "ace list" (the output of machu.2.ace()) can both be provided. Any
combination of taxa can be displayed either in separate plots or on the same plot (using matplot()),
and any combination of climate variables used in constructing the models can also be displayed. A
few arguments for adjusting line width, color, etc. are provided.
}
