% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machuruku_v0.1.8.R
\name{machu.tree.unc}
\alias{machu.tree.unc}
\title{Account for uncertainty in divergence time estimation from a single tree}
\usage{
machu.tree.unc(tree)
}
\arguments{
\item{tree}{input treedata object, imported with the read.beast() function from treeio.
Only tested with output from the BEAST 2 utility TreeAnnotator.}
}
\value{
returns a list of three treedata objects better characterizing the uncertainty
of the input tree's node heights.
}
\description{
This function generates two additional trees from a single input tree with uncertainty
(e.g. 95\% confidence intervals) in divergence time estimations. This function is written to
work with output from the BEAST 2 program TreeAnnotator and may not work with other formats.
One tree will be built using the lower confidence interval values for each node (i.e. a
more recent tree), and the other tree will be built using the upper confidence interval
values for each node (i.e. an older tree). The output of the function is a list of three
treedata objects, each of which can be separately used as tree input for Machuruku.
As the values of 95\% confidence intervals do not generally make an ultrametric tree when
used as node heights, the phytools function force.ultrametric() is used to force the
two generated trees to be ultrametric using the "nnls" function, which may result in
some shifting of node heights from the 95\% CI values.
}
\examples{
trees <- machu.tree.unc(beast)
}
