% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machuruku_v0.1.8.3.R
\name{machu.treeplot}
\alias{machu.treeplot}
\title{Plot trees to visualize each step in the Machuruku process}
\usage{
machu.treeplot(
  tree,
  upperX = 20,
  nodecirclecolor = "cadetblue1",
  nodecirclesize = 7,
  nodelabelsize = 5,
  tiplabelsize = 5,
  timelabeloffset = 3,
  timelabelsize = 5,
  timelabeldigits = 2,
  timeslice = NULL,
  timeslicecolor = "coral1",
  timeslicelinetype = "solid"
)
}
\arguments{
\item{tree}{tree input. Can be a single tree in treedata or phylo format, or a list of trees in either
format. Output of the plotting function will change depending on the type of input.}

\item{upperX}{set the upper (right-side) limit of the timescale for the treeplot. Must be a
positive number. The real purpose of this value is to scale the treeplot so that the tip labels
are visible. At a value of 0 the tip labels will likely be cut off by the right border of the plot.}

\item{nodecirclecolor}{set the color for the node circles. Set to "cadetblue1" by default.}

\item{nodecirclesize}{set the size for the node circles. Set to 7 by default.}

\item{nodelabelsize}{set the size for the node label text (node numbers). Set to 5 by default.}

\item{tiplabelsize}{set the size for the tip label text. Set to 5 by default.}

\item{timelabeloffset}{set the horizontal offset for the divergence time labels. At different scales
the tree plot will place the divergence time labels at varying distances from their corresponding
nodes. This value shifts them horizontally so that the user can make an aesthetically pleasing image
at their preferred scale.}

\item{timelabelsize}{set the size for the divergence time label text. Set to 5 by default.}

\item{timelabeldigits}{set the number of node label (divergence time) decimal points displayed. Default is 2.}

\item{timeslice}{display a vertical line through the phylogeny at the desired time-slice for a
machu.2.ace analysis. By default this feature is turned off.}

\item{timeslicecolor}{set the color for the time-slice line. Set to "coral1" by default.}

\item{timeslicelinetype}{set the line-type for the time-slice line. Set to "solid" by default. Other
options include "dashed" and "dotted".}
}
\value{
Returns a tree plot with different properties depending on whether the input is a single tree
or a list of trees from the functions machu.tree.unc() and machu.trees.unc().
}
\description{
Plot trees with node labels and divergence dates on a time scale. Accepts trees as a
single "phylo" or "treedata" object, or a list of "treedata" objects (output from the
machu.tree.unc() or machu.trees.unc() functions). Visualizing a single tree allows for the
assessment of the proper value for T encompassing the desired number of lineages, and
shows node numbers so that the name of a given lineage (which for an internal branch may
be something like node5-node6) can be easily attributed to a particular branch by the user.
Visualizing a list of trees on the same timescale visualizes the different branch-taxa recoverable
by a single time-slice for trees with multiple heights.
}
\examples{
# plot three trees derived from machu.tree.unc (a list of three treedata objects)
machu.treeplot(trees, upperX=3, timelabeloffset=1.25, nodecirclecolor="orangered1")
# plot three trees derived from machu.tree.unc with a timeslice at 2.5 mya
machu.treeplot(trees, upperX=3, timelabeloffset=1.25, nodecirclecolor="orangered1", timeslice = 2.5)
# plot a single treedata object with no time-slice
machu.treeplot(tree, upperX=3, nodecirclecolor = "lightblue", timelabeloffset = 1)
# plot a single treedata object with time-slice at 3 mya
machu.treeplot(tree, upperX=3, nodecirclecolor = "lightblue", timelabeloffset = 1, timeslice=3, timeslicecolor = "blue", timeslicelinetype = "dashed")
}
